--- src/components/SideBar/index.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-type Props = {};
-
-export function SideBar({}: Props) {
-  return <div>Sidebar</div>;
-}
