+++ src/components/TodoList/index.tsx
@@ -6,8 +6,13 @@ export function TodoList() {
   const { tasks } = useDataProvider();
   return (
     <Styled.TodoList>
-      {tasks.length > 0 &&
-        tasks.map((task) => <TodoListItem task={task} key={task._id} />)}
+      {tasks.length > 0 ? (
+        tasks.map((task) => <TodoListItem task={task} key={task._id} />)
+      ) : (
+        <Styled.EmptyTodoList>
+          There's nothing in TO-DO, add something using + at the bottom right
+        </Styled.EmptyTodoList>
+      )}
     </Styled.TodoList>
   );
 }
