+++ src/routes/Home/index.tsx
@@ -7,6 +7,7 @@ import {
 } from "../../components";
 import { RiAddLine } from "react-icons/ri";
 import { useDataProvider } from "../../context/data-context";
+import { useDocumentTitle } from "../../hooks/useDocumentTitle";
 
 const FullScreen = styled.div`
   height: 100vh;
@@ -24,6 +25,8 @@ export const Home = () => {
   const { modal, dispatch } = useDataProvider();
   const createTask = () =>
     dispatch({ type: "TOGGLE_MODAL", payload: { visible: true, edit: false } });
+
+  useDocumentTitle("Pomotodoro | Home");
 
   return (
     <FullScreen>
